refactor migrations
