+++ backside/.env.example
@@ -1,4 +1,10 @@
 SK8_ADDRESS=
 SK8_PORT=8080
 
+SK8_DB_HOST=localhost
+SK8_DB_PORT=5432
+SK8_DB_USER=root
+SK8_DB_PASSWORD=root
+SK8_DB_NAME=sk8
+
 SK8_PLACES_TOMTOM_API_KEY=<your_tomtom_api_key>
