+++ backside/.env.testing
@@ -1,7 +1,7 @@
 SK8_PLACES_TOMTOM_API_KEY=fake
 
-SK8_SPOTS_DB_HOST=localhost
-SK8_SPOTS_DB_PORT=5432
-SK8_SPOTS_DB_USER=root
-SK8_SPOTS_DB_PASSWORD=root
-SK8_SPOTS_DB_NAME=spots
+SK8_DB_HOST=fake
+SK8_DB_PORT=1234
+SK8_DB_USER=fake
+SK8_DB_PASSWORD=fake
+SK8_DB_NAME=fake
