+++ backside/Makefile
@@ -1,2 +1,1 @@
-clean:
-	go run spots/scripts/drop_spots_db_tables.go
+# TODO
